Python_Web_1/requirements.txt
cffi==0.9.2
cryptography==0.8.2
enum34==1.0.4
Flask==0.10.1
Flask-Admin==1.1.0
Flask-Bcrypt==0.6.2
Flask-Login==0.2.11
Flask-MySQL==1.2
flask-paginate==0.2.8
Flask-SQLAlchemy==2.0
Flask-Testing==0.4.2
Flask-WTF==0.11
gnureadline==6.3.3
ipython==2.4.1
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
MySQL-python==1.2.5
pyasn1==0.1.7
pycparser==2.12
pycrypto==2.6.1
pysha3==0.3
python-bcrypt==0.3.1
six==1.9.0
SQLAlchemy==0.9.8
Werkzeug==0.10.1
WTForms==2.0.2
